🔧 App Router構成を最適化してサーバーコンポーネントとクライアントコンポーネントを適切に分離
- layout.tsxをサーバーコンポーネントに変更
- 全体をラップするClientThemeWrapperとprovidersを削除
- MUIが必要な個別コンポーネント用のMuiProviderを作成
- DashboardWithCalendarのみMuiProviderでラップ
- サインインページなどMUI不要のページはサーバーコンポーネントとして動作
- App Routerの利点（SEO、初期読み込み速度）を最大化

Fixes #34
